+++ Entregas/Assignement_1/multMatrixOnHost.h
@@ -1,7 +1,7 @@
-void multMatrixOnHost(float *A, float *B, float *C, const int nx, const int ny){
-  float *ia = A;
-  float *ib = B;
-  float *ic = C;
+void multMatrixOnHost(int *A, int *B, int *C, const int nx, const int ny){
+  int *ia = A;
+  int *ib = B;
+  int *ic = C;
 
   for (int i = 0; i < ny; i++) {
     for (int j = 0; j < nx; j++) {
